Added wall property "ignore marker", allowing doors to close even if a marker sits in them.
git-svn-id: https://d2x-xl.svn.sourceforge.net/svnroot/d2x-xl@4993 01821946-851b-0410-b7f4-f1477c70e57a

+++ dle-xp/types.h
@@ -540,7 +540,7 @@ public:
   FIX   hps;            /* "Hit points" of the wall.  */
   INT32 linked_wall;    /* number of linked wall */
   UINT8 type;           /* What kind of special wall. */
-  UINT8 flags;          /* Flags for the wall.     */
+  UINT16 flags;          /* Flags for the wall.     */
   UINT8 state;          /* Opening, closing, etc. */
   UINT8 trigger;        /* Which trigger is associated with the wall. */
   INT8  clip_num;       /* Which  animation associated with the wall.  */
